PrismaClient 브라우저 실행 오류 수정 및 서버 사이드 전용 처리
