+++ app/api/tweets/route.ts
@@ -1,0 +1,7 @@
+import db from '@/lib/db';
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  const tweets = await db.tweet.findMany({ include: { user: true, like: true } });
+  return NextResponse.json(tweets);
+}
