+++ package.json
@@ -10,6 +10,9 @@
   },
   "dependencies": {
     "@prisma/client": "^6.6.0",
+    "@types/bcrypt": "^5.0.2",
+    "bcrypt": "^5.1.1",
+    "iron-session": "^8.0.4",
     "next": "15.3.1",
     "prisma": "^6.6.0",
     "react": "^19.0.0",
